pyproject.toml
[project]
name = "regional-dd-wm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "anemoi-datasets==0.5.7",
    "anemoi-graphs==0.3.0",
    "anemoi-inference[huggingface]==0.4.9",
    "anemoi-models==0.3.1",
    "anemoi-transform==0.1.16",
    "boto3==1.40.32",
    "earthkit==0.12.0",
    "earthkit-regrid==0.4.0",
    "ecmwf-opendata==0.3.23",
    "huggingface-hub==0.35.0",
    "ipykernel==6.30.1",
    "marimo==0.15.5",
    "matplotlib==3.10.6",
    "numpy==1.26.4",
    # install flash-attn only on Linux (CUDA-capable systems)
    # "flash-attn; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "anemoi-registry>=0.1.8",
    "cartopy>=0.25.0",
]

[tool.uv.extra-build-dependencies]
flash-attn = ["setuptools", "torch"]
